+++ Sources/SuperwallKit/Analytics/Internal Tracking/Trackable Events/TrackableSuperwallPlacement.swift
@@ -194,12 +194,6 @@ enum InternalSuperwallPlacement {
     func getSuperwallParameters() async -> [String: Any] { [:] }
   }
 
-  struct ConfigRefresh: TrackableSuperwallPlacement {
-    let superwallPlacement: SuperwallPlacement = .configRefresh
-    var audienceFilterParams: [String: Any] = [:]
-    func getSuperwallParameters() async -> [String: Any] { [:] }
-  }
-
   struct ConfigAttributes: TrackableSuperwallPlacement {
     let superwallPlacement: SuperwallPlacement = .configAttributes
     let options: SuperwallOptions
@@ -600,6 +594,12 @@ enum InternalSuperwallPlacement {
     }
   }
 
+  struct ConfirmAllAssignments: TrackableSuperwallPlacement {
+    let superwallPlacement: SuperwallPlacement = .confirmAllAssignments
+    let audienceFilterParams: [String: Any] = [:]
+    func getSuperwallParameters() async -> [String: Any] { [:] }
+  }
+
   struct FreeTrialStart: TrackableSuperwallPlacement {
     var superwallPlacement: SuperwallPlacement {
       return .freeTrialStart(
@@ -718,6 +718,41 @@ enum InternalSuperwallPlacement {
       }
       params += await paywallInfo.placementParams()
       return params
+    }
+  }
+
+  enum ConfigCacheStatus: String {
+    case cached = "CACHED"
+    case notCached = "NOT_CACHED"
+  }
+
+  struct ConfigRefresh: TrackableSuperwallPlacement {
+    let superwallPlacement: SuperwallPlacement = .configRefresh
+    let buildId: String
+    let retryCount: Int
+    let cacheStatus: ConfigCacheStatus
+    let fetchDuration: TimeInterval
+    var audienceFilterParams: [String: Any] = [:]
+
+    func getSuperwallParameters() async -> [String: Any] {
+      return [
+        "config_build_id": buildId,
+        "retry_count": retryCount,
+        "cache_status": cacheStatus.rawValue,
+        "fetch_duration": fetchDuration
+      ]
+    }
+  }
+
+  struct ConfigFail: TrackableSuperwallPlacement {
+    let superwallPlacement: SuperwallPlacement = .configFail
+    let message: String
+    var audienceFilterParams: [String: Any] = [:]
+
+    func getSuperwallParameters() async -> [String: Any] {
+      return [
+        "error_message": message
+      ]
     }
   }
 }