+++ Sources/SuperwallKit/Storage/Cache/CacheKeys.swift
@@ -160,3 +160,19 @@ enum DisableVerbosePlacements: Storable {
   static var directory: SearchPathDirectory = .appSpecificDocuments
   typealias Value = Bool
 }
+
+enum LatestConfig: Storable {
+  static var key: String {
+    "store.config"
+  }
+  static var directory: SearchPathDirectory = .appSpecificDocuments
+  typealias Value = Config
+}
+
+enum LatestGeoInfo: Storable {
+  static var key: String {
+    "store.geoInfo"
+  }
+  static var directory: SearchPathDirectory = .appSpecificDocuments
+  typealias Value = GeoInfo
+}