+++ Tests/SuperwallKitTests/Network/NetworkMock.swift
@@ -23,6 +23,7 @@ final class NetworkMock: Network {
   @MainActor
   override func getConfig(
     injectedApplicationStatePublisher: (AnyPublisher<UIApplication.State, Never>)? = nil,
+    maxRetry: Int? = nil,
     isRetryingCallback: ((Int) -> Void)? = nil
   ) async throws -> Config {
     getConfigCalled = true