+++ Tests/SuperwallKitTests/Paywall/View Controller/Web View/Message Handling/PaywallMessageHandlerDelegateMock.swift
@@ -10,9 +10,16 @@ import Foundation
 
 final class FakeWebView: SWWebView {
   var willHandleJs = false
-  override func evaluateJavaScript(_ javaScriptString: String, completionHandler: ((Any?, Error?) -> Void)? = nil) {
+
+  #if compiler(>=6.0.0)
+  override func evaluateJavaScript(_ javaScriptString: String, completionHandler: (@MainActor (Any?, (any Error)?) -> Void)? = nil) {
     willHandleJs = true
   }
+  #else
+  override func evaluateJavaScript(_ javaScriptString: String, completionHandler: ((Any?, (any Error)?) -> Void)? = nil) {
+    willHandleJs = true
+  }
+  #endif
 }
 
 final class PaywallMessageHandlerDelegateMock: PaywallMessageHandlerDelegate {